+++ src/test/groovy/com/ullink/gradle/symstore/AddSymbolTaskTest.groovy
@@ -23,13 +23,32 @@ public class AddSymbolTaskTest {
     @Test
     void 'Given echo command when running the task then parameters should be run without exception.'() {
         AddSymbolTask task = ApplyPluginAndGetTask('echo %*')
+        def share = File.createTempDir()
+        task.share = share
+        task.file = new File('the_file')
+        task.product = 'the_product'
+        task.run()
+        share.deleteDir()
+    }
+
+    @Test
+    void 'Given echo command and no share when running the task then exception should be thrown.'() {
+        AddSymbolTask task = ApplyPluginAndGetTask('echo %*')
+        task.file = new File('the_file')
+        task.product = 'the_product'
+        exception.expect(GradleException)
         task.run()
     }
 
     @Test
     void 'Given "exit 1" command when running the task then exception should be thrown.'() {
         AddSymbolTask task = ApplyPluginAndGetTask('exit 1')
+        def share = File.createTempDir()
+        task.share = share
+        task.file = new File('the_file')
+        task.product = 'the_product'
         exception.expect(GradleException)
         task.run()
+        share.deleteDir()
     }
 }