+++ CHANGELOG.md
@@ -4,5 +4,6 @@ Plugin changelog
 0.1
 -------
 
+* Added lock for share folder
 * Added "addSymbol" task to add symbols to a store.